requirements.txt
aioauth-client
aiohttp
aiohttp_jinja2
jinja2
jsonschema
motor
pytest
pyyaml
